Convert output to callgrind format.
Enable the output to be viewed in KCachegrind.
Compress the stacks by removing internal Python calls and everything
above Py_Main.
Add 'vmprofrun' to automate collection of data and conversion
to callgrind formation.
Add 'vmproftocallgrind' for manual conversion.
Expand the maximum function name length in the collector so the entire path
to the python file can be collected.
The --perf option was added to vmprofshow, but using it to add the perf map file
doesn't affect the output.