Improve symbol resolution for JIT functions.
Sort the symbols for the JIT function so the addrspace lookup function
works properly.

Add the '<JIT>' name to the library (vs. the previous use of None to
indicated a JIT'ed library). This makes it clearer when a JIT'ed
function is handled vs. some address in a unknown library.

Add an option to remove the Numba Dispatcher_call from the callstack.
The full stacks know which python code calls which compiled function,
but the callgrind format does not.  So every call to a compiled function
goes through 'Dispatcher_call', and then back out the compiled function.
Using this option will keep the stacks separated in the GUI.